+++ sandwich-contexts-kubernetes/lib/Test/Sandwich/Contexts/Kubernetes/KataContainers.hs
@@ -151,7 +151,7 @@ withKataContainers' kcc@(KubernetesClusterContext {..}) kubectlBinary options@(K
 
   info [i|kataRoot: #{kataRoot}|]
 
-  env <- askKubectlEnvironment kcc
+  env <- getKubectlEnvironment kcc
 
   -- Now follow the instructions from
   -- https://github.com/kata-containers/kata-containers/blob/main/docs/install/minikube-installation-guide.md#installing-kata-containers