+++ sandwich-contexts-kubernetes/lib/Test/Sandwich/Contexts/Kubernetes/Namespace.hs
@@ -11,9 +11,18 @@ Helper module for working with Kubernetes namespaces.
 module Test.Sandwich.Contexts.Kubernetes.Namespace (
   withKubernetesNamespace
   , withKubernetesNamespace'
+  , withKubernetesNamespace''
+  , withKubernetesNamespace'''
 
+  -- * Create a namespace
   , createKubernetesNamespace
+  , createKubernetesNamespace'
+  , createKubernetesNamespace''
+
+  -- * Destroy a namespace
   , destroyKubernetesNamespace
+  , destroyKubernetesNamespace'
+  , destroyKubernetesNamespace''
   ) where
 
 import Control.Monad
@@ -21,6 +30,7 @@ import Data.String.Interpolate
 import Relude hiding (force)
 import System.Exit
 import Test.Sandwich
+import Test.Sandwich.Contexts.Files
 import Test.Sandwich.Contexts.Kubernetes.Kubectl
 import Test.Sandwich.Contexts.Kubernetes.Types
 import UnliftIO.Exception
@@ -48,7 +58,36 @@ withKubernetesNamespace' :: (
   => Text
   -> m a
   -> m a
-withKubernetesNamespace' namespace = bracket_ (createKubernetesNamespace namespace) (destroyKubernetesNamespace False namespace)
+withKubernetesNamespace' namespace =
+  bracket_ (createKubernetesNamespace namespace) (destroyKubernetesNamespace False namespace)
+
+-- | Same as 'withKubernetesNamespace'', but allows you to pass in the path to the @kubectl@ binary.
+withKubernetesNamespace'' :: (
+  KubernetesClusterBasic context m
+  )
+  -- | Path to @kubectl@ binary
+  => FilePath
+  -- | Namespace to create
+  -> Text
+  -> m a
+  -> m a
+withKubernetesNamespace'' kubectl namespace =
+  bracket_ (createKubernetesNamespace' kubectl namespace) (destroyKubernetesNamespace' kubectl False namespace)
+
+-- | Same as 'withKubernetesNamespace''', but allows you to pass in the path to the cluster context.
+withKubernetesNamespace''' :: (
+  KubernetesClusterBasic context m
+  )
+  -- | Cluster context
+  => KubernetesClusterContext
+  -- | Path to @kubectl@ binary
+  -> FilePath
+  -- | Namespace to create
+  -> Text
+  -> m a
+  -> m a
+withKubernetesNamespace''' kcc kubectl namespace =
+  bracket_ (createKubernetesNamespace'' kcc kubectl namespace) (destroyKubernetesNamespace'' kcc kubectl False namespace)
 
 -- | Create a Kubernetes namespace.
 createKubernetesNamespace :: (
@@ -57,11 +96,20 @@ createKubernetesNamespace :: (
   -- | Namespace name
   => Text
   -> m ()
-createKubernetesNamespace namespace = do
-  let args = ["create", "namespace", toString namespace]
-  (kubectl, env) <- askKubectlArgs
-  createProcessWithLogging ((proc kubectl args) { env = Just env, delegate_ctlc = True })
-    >>= waitForProcess >>= (`shouldBe` ExitSuccess)
+createKubernetesNamespace namespace =
+  askFile @"kubectl" >>= flip createKubernetesNamespace' namespace
+
+-- | Create a Kubernetes namespace.
+createKubernetesNamespace' :: (
+  KubernetesClusterBasic context m
+  )
+  -- | Path to @kubectl@ binary
+  => FilePath
+  -- | Namespace name
+  -> Text
+  -> m ()
+createKubernetesNamespace' kubectl namespace =
+  getContext kubernetesCluster >>= (\kcc -> createKubernetesNamespace'' kcc kubectl namespace)
 
 -- | Destroy a Kubernetes namespace.
 destroyKubernetesNamespace :: (
@@ -72,9 +120,56 @@ destroyKubernetesNamespace :: (
   -- | Namespace name
   -> Text
   -> m ()
-destroyKubernetesNamespace force namespace = do
+destroyKubernetesNamespace force namespace =
+  askFile @"kubectl" >>= (\x -> destroyKubernetesNamespace' x force namespace)
+
+-- | Destroy a Kubernetes namespace.
+destroyKubernetesNamespace' :: (
+  KubernetesClusterBasic context m
+  )
+  -- | Path to @kubectl@ binary
+  => FilePath
+  -- | Whether to pass @--force@
+  -> Bool
+  -- | Namespace name
+  -> Text
+  -> m ()
+destroyKubernetesNamespace' kubectl force namespace = do
+  getContext kubernetesCluster >>= (\kcc -> destroyKubernetesNamespace'' kcc kubectl force namespace)
+
+-- | Create a Kubernetes namespace.
+createKubernetesNamespace'' :: (
+  KubernetesClusterBasic context m
+  )
+  -- | Cluster context
+  => KubernetesClusterContext
+  -- | Path to @kubectl@ binary
+  -> FilePath
+  -- | Namespace name
+  -> Text
+  -> m ()
+createKubernetesNamespace'' kcc kubectl namespace = do
+  let args = ["create", "namespace", toString namespace]
+  env <- getKubectlEnvironment kcc
+  createProcessWithLogging ((proc kubectl args) { env = Just env, delegate_ctlc = True })
+    >>= waitForProcess >>= (`shouldBe` ExitSuccess)
+
+-- | Destroy a Kubernetes namespace.
+destroyKubernetesNamespace'' :: (
+  KubernetesClusterBasic context m
+  )
+  -- | Cluster context
+  => KubernetesClusterContext
+  -- | Path to @kubectl@ binary
+  -> FilePath
+  -- | Whether to pass @--force@
+  -> Bool
+  -- | Namespace name
+  -> Text
+  -> m ()
+destroyKubernetesNamespace'' kcc kubectl force namespace = do
   let args = ["delete", "namespace", toString namespace]
            <> if force then ["--force"] else []
-  (kubectl, env) <- askKubectlArgs
+  env <- getKubectlEnvironment kcc
   createProcessWithLogging ((proc kubectl args) { env = Just env, delegate_ctlc = True })
     >>= waitForProcess >>= (`shouldBe` ExitSuccess)